refactor: Refactored constraint namespace. | TA-2792 #time 20m
