+++ src/TenantCloud/APIVersioning/APIVersioningServiceProvider.php
@@ -6,10 +6,10 @@ use Illuminate\Http\Request;
 use Illuminate\Routing\Contracts\ControllerDispatcher;
 use Illuminate\Routing\Route;
 use Illuminate\Support\ServiceProvider;
-use TenantCloud\APIVersioning\Constrain\BuiltInStringConstraintParser;
-use TenantCloud\APIVersioning\Constrain\ConstraintChecker;
-use TenantCloud\APIVersioning\Constrain\SemanticConstraintChecker;
-use TenantCloud\APIVersioning\Constrain\StringConstraintParser;
+use TenantCloud\APIVersioning\Constraint\BuiltInStringConstraintParser;
+use TenantCloud\APIVersioning\Constraint\ConstraintChecker;
+use TenantCloud\APIVersioning\Constraint\SemanticConstraintChecker;
+use TenantCloud\APIVersioning\Constraint\StringConstraintParser;
 use TenantCloud\APIVersioning\Version\RequestHeaderVersionParser;
 use TenantCloud\APIVersioning\Version\RequestVersionParser;
 use TenantCloud\APIVersioning\Version\SemanticVersionParser;
