+++ src/TenantCloud/APIVersioning/Constraint/StringConstraintParser.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TenantCloud\APIVersioning\Constrain;
+namespace TenantCloud\APIVersioning\Constraint;
 
 interface StringConstraintParser
 {
